refactoring into a better boilerplate structure. functionality is still missing but this could be more sustainable long term
